Validation checks
Unfortunately error handling is very inconsistent between SAX and
DOMParser. Best way to secure any errors is through preventing JS errors
in Anti. We'll eventually produce an output for bad and good inputs that
will be safe to use. The purpose of Anti is not HTML/XML validation so
I'll leave any hardcore attempts